Moving 2 tests from KafkaMP to KafkaSE
